use int64

+++ go/client/user_service/user_service.go
@@ -99,7 +99,7 @@ type CreatedSession struct {
 	FirstConnectedAt *time.Time `json:"first_connected_at,omitempty"`
 	Label            *string    `json:"label"`
 	Node             *string    `json:"node"`
-	PeerId           *int       `json:"peer_id,omitempty"`
+	PeerId           *int64     `json:"peer_id,omitempty"`
 	ToDelete         *bool      `json:"to_delete,omitempty"`
 	TokenId          *string    `json:"token_id"`
 }
@@ -206,7 +206,7 @@ type PatchedSession struct {
 	FirstConnectedAt *time.Time `json:"first_connected_at,omitempty"`
 	Label            *string    `json:"label,omitempty"`
 	Node             *string    `json:"node,omitempty"`
-	PeerId           *int       `json:"peer_id,omitempty"`
+	PeerId           *int64     `json:"peer_id,omitempty"`
 	ToDelete         *bool      `json:"to_delete,omitempty"`
 	TokenId          *string    `json:"token_id,omitempty"`
 	UpdatedAt        *time.Time `json:"updated_at"`
@@ -223,7 +223,7 @@ type PatchedUser struct {
 // PeerUser defines model for PeerUser.
 type PeerUser struct {
 	Params    map[string]interface{} `json:"params"`
-	PeerId    int                    `json:"peer_id"`
+	PeerId    int64                  `json:"peer_id"`
 	PeerLabel string                 `json:"peer_label"`
 	QueryId   string                 `json:"query_id"`
 }
@@ -247,7 +247,7 @@ type Session struct {
 	Id               *string    `json:"id,omitempty"`
 	Label            *string    `json:"label,omitempty"`
 	Node             *string    `json:"node,omitempty"`
-	PeerId           *int       `json:"peer_id,omitempty"`
+	PeerId           *int64     `json:"peer_id,omitempty"`
 	ToDelete         *bool      `json:"to_delete,omitempty"`
 	TokenId          *string    `json:"token_id,omitempty"`
 	UpdatedAt        *time.Time `json:"updated_at,omitempty"`
@@ -311,7 +311,7 @@ type UpdatedSession struct {
 	FirstConnectedAt *time.Time `json:"first_connected_at"`
 	Label            *string    `json:"label"`
 	Node             *string    `json:"node"`
-	PeerId           *int       `json:"peer_id"`
+	PeerId           *int64     `json:"peer_id"`
 	ToDelete         *bool      `json:"to_delete"`
 	TokenId          *string    `json:"token_id"`
 	UpdatedAt        *time.Time `json:"updated_at"`
